refactor : Update media split router
