+++ server/internal/application/user_handler.go
@@ -2,7 +2,6 @@ package application
 
 import (
 	"dalkak/internal/infrastructure/eventbus"
-	mediadto "dalkak/pkg/dto/media"
 	userdto "dalkak/pkg/dto/user"
 	cryptoutil "dalkak/pkg/utils/crypto"
 	jwtutil "dalkak/pkg/utils/jwt"
@@ -13,7 +12,6 @@ import (
 func (app *ApplicationImpl) RegisterUserEventListeners() {
 	app.EventManager.Subscribe("post.user.auth", app.handleAuthAndSignUp)
 	app.EventManager.Subscribe("post.user.refresh", app.handleReissueAccessToken)
-	app.EventManager.Subscribe("post.user.media.presigned", app.handleCreateTempMedia)
 }
 
 func (app *ApplicationImpl) handleAuthAndSignUp(event eventbus.Event) {
@@ -84,37 +82,5 @@ func (app *ApplicationImpl) handleReissueAccessToken(event eventbus.Event) {
 
 	// 리턴
 	result := userdto.NewReissueAccessTokenResponse(accessToken.Token, accessToken.TokenTTL)
-	app.SendResponse(event.ResponseChan, responseutil.NewAppData(result, responseutil.DataCodeCreated), nil)
-}
-
-func (app *ApplicationImpl) handleCreateTempMedia(event eventbus.Event) {
-	userInfo := event.UserInfo
-	if userInfo == nil {
-		app.SendResponse(event.ResponseChan, nil, responseutil.NewAppError(responseutil.ErrCodeUnauthorized, responseutil.ErrMsgRequestUnauth))
-		return
-	}
-	payload, ok := event.Payload.(*userdto.CreateTempMediaRequest)
-	if !ok {
-		app.SendResponse(event.ResponseChan, nil, responseutil.NewAppError(responseutil.ErrCodeBadRequest, responseutil.ErrMsgRequestInvalid))
-		return
-	}
-
-	// 미디어 생성
-	dto := mediadto.NewCreateTempMediaDto(userInfo, payload.MediaType, payload.Ext, payload.Prefix)
-	newMedia, err := app.MediaDomain.CreateMediaTemp(dto)
-	if err != nil {
-		app.SendResponse(event.ResponseChan, nil, err)
-		return
-	}
-
-	// 미디어 저장
-	err = app.Database.CreateUserMediaTemp(userInfo.GetUserId(), newMedia)
-	if err != nil {
-		app.SendResponse(event.ResponseChan, nil, err)
-		return
-	}
-
-	// 리턴
-	result := userdto.NewUserCreateMediaResponse(newMedia.MediaEntity.Id, newMedia.MediaTempUrl.AccessUrl, *newMedia.MediaTempUrl.UploadUrl)
 	app.SendResponse(event.ResponseChan, responseutil.NewAppData(result, responseutil.DataCodeCreated), nil)
 }
