+++ server/internal/infrastructure/web/user_router.go
@@ -36,19 +36,4 @@ func SetupUserRoute(group fiber.Router, keyManager core.KeyManager, eventManager
 
 		return PublishAndWaitResponse(eventManager, "post.user.refresh", nil, req)
 	}))
-
-	group.Post("/media/presigned", WarpHandler(func(c fiber.Ctx) interface{} {
-		user, err := GetUserInfoFromContext(c, true)
-		if err != nil {
-			return err
-		}
-
-		req := new(userdto.CreateTempMediaRequest)
-		err = BindAndValidate(c, req)
-		if err != nil {
-			return err
-		}
-
-		return PublishAndWaitResponse(eventManager, "post.user.media.presigned", user, req)
-	}))
 }
